+++ TankGame/src/Mesh.cpp
@@ -681,7 +681,7 @@ std::unique_ptr<Mesh> OBJLoader::LoadOBJ(const std::string & filePath)
 	std::vector<faceDef> faces;
 
 	std::ifstream file(filePath, std::ios::in | std::ios::ate);
-	long fileLength = file.tellg();
+	long long fileLength = file.tellg();
 	file.seekg(0, std::ios::beg);
 	fileLength -= file.tellg();
